San Ceriz -> Saint Cerise
so I did some searches and found that "Saint Arkh" starts with "sento" in Japanese, as expected, but "Saint-Gral" and "Saint-Croix" start with "san", like the city here.

The only instance of "San" I could find in English is the store named "San Corries" in Cold Steel, which also starts with "san" in Japanese. But then they use the English word "corries", which comes from Gaelic, and not Spanish like "Correas" or something

anyway, since we're anglicizing stuff, might as well go back to "Cerise", the only explicitly English word I considered.